travis-ci bug causing failure on ppc64le
Would you be interested to try running the build in debug mode? With access to the VM and the intimate understanding of your tools and code base, you are in the best position to troubleshoot this issue further. The debug mode is explained in https://docs.travis-ci.com/user/running-build-in-debug-mode/.

Debug mode for the repository ai-techsystems/dnnCompiler has been enabled.